Moved tuple_with_vec tests to tuple_with_vec_int
