+++ xsd-parser/tests/tuple_with_vec_int/example.xml
